audio_transcriber/requirements.txt
fastmcp
python-dotenv
openai-whisper
